+++ src/Scripting/Postponer.ts
@@ -1,5 +1,5 @@
 import type { Moment, unitOfTime } from 'moment';
-import { DateFallback } from '../DateFallback';
+import { DateFallback } from '../Task/DateFallback';
 import { Task } from '../Task/Task';
 import { TasksDate } from './TasksDate';
