+++ src/Task/Task.ts
@@ -3,7 +3,6 @@ import { getSettings, getUserSelectedTaskFormat } from '../Config/Settings';
 import { GlobalFilter } from '../Config/GlobalFilter';
 import { StatusRegistry } from '../Statuses/StatusRegistry';
 import type { Status } from '../Statuses/Status';
-import { DateFallback } from '../DateFallback';
 import { compareByDate } from '../lib/DateTools';
 import { TasksDate } from '../Scripting/TasksDate';
 import { StatusType } from '../Statuses/StatusConfiguration';
@@ -11,6 +10,7 @@ import { TasksFile } from '../Scripting/TasksFile';
 import { PriorityTools } from '../lib/PriorityTools';
 import { logging } from '../lib/logging';
 import { logEndOfTaskEdit, logStartOfTaskEdit } from '../lib/LogTasksHelper';
+import { DateFallback } from './DateFallback';
 import { Urgency } from './Urgency';
 import type { Recurrence } from './Recurrence';
 import type { TaskLocation } from './TaskLocation';
