+++ tests/Task/DateFallback.test.ts
@@ -3,11 +3,11 @@
  */
 import type { Moment } from 'moment';
 import moment from 'moment';
-import { Task } from '../src/Task/Task';
-import { resetSettings, updateSettings } from '../src/Config/Settings';
-import { DateFallback } from '../src/DateFallback';
-import { TaskLocation } from '../src/Task/TaskLocation';
-import { TaskBuilder } from './TestingTools/TaskBuilder';
+import { Task } from '../../src/Task/Task';
+import { resetSettings, updateSettings } from '../../src/Config/Settings';
+import { DateFallback } from '../../src/Task/DateFallback';
+import { TaskLocation } from '../../src/Task/TaskLocation';
+import { TaskBuilder } from '../TestingTools/TaskBuilder';
 
 jest.mock('obsidian');
 window.moment = moment;
